+++ features/tickets/domain/src/main/java/ru/sulgik/tickets/domain/entity/Schedule.kt
@@ -23,19 +23,21 @@ data class Schedule(
     )
 
     data class Place(
-        val price: Double,
+        val price: Int,
         val type: FilmHallCellType,
     )
 
     data class PayedTicket(
-        val column: Double,
+        val column: Int,
         val filmId: String,
         val phone: String,
-        val row: Double,
+        val row: Int,
     )
 
     enum class FilmHallCellType {
         BLOCKED, COMPORT, ECONOM
     }
+
+
 
 }
